requirements.txt
# Core (Python 3.10–3.12 recommended; TensorFlow may not support 3.13+ yet — see README)
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0
scipy>=1.11.0

# Data
yfinance>=0.2.40

# ML
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost>=2.0.0
# LSTM: install when your Python version supports wheel (often 3.10–3.12)
tensorflow>=2.15.0; python_version < "3.13"

# Volatility (GARCH)
arch>=6.2.0

# NLP / Sentiment
nltk>=3.8.0
vaderSentiment>=3.3.2
transformers>=4.36.0
torch>=2.1.0

# Optional Twitter (set TWITTER_BEARER_TOKEN)
tweepy>=4.14.0

# Explainability
shap>=0.44.0

# App & viz
streamlit>=1.29.0
plotly>=5.18.0
matplotlib>=3.8.0

# Config
pyyaml>=6.0.0

# Experiment tracking
mlflow>=2.10.0
